+++ core/src/utils/menu-controller/index.ts
@@ -1,8 +1,10 @@
+import { doc } from '@utils/browser';
+import type { BackButtonEvent } from '@utils/hardware-back-button';
+import { MENU_BACK_BUTTON_PRIORITY } from '@utils/hardware-back-button';
 import { printIonWarning } from '@utils/logging';
 
 import type { MenuI, MenuControllerI } from '../../components/menu/menu-interface';
-import type { AnimationBuilder, BackButtonEvent } from '../../interface';
-import { MENU_BACK_BUTTON_PRIORITY } from '../hardware-back-button';
+import type { AnimationBuilder } from '../../interface';
 import { componentOnReady } from '../helpers';
 
 import { menuOverlayAnimation } from './animations/overlay';
@@ -227,17 +229,14 @@ const createMenuController = (): MenuControllerI => {
   registerAnimation('push', menuPushAnimation);
   registerAnimation('overlay', menuOverlayAnimation);
 
-  if (typeof document !== 'undefined') {
-    document.addEventListener('ionBackButton', (ev: any) => {
-      // TODO(FW-2832): type
-      const openMenu = _getOpenSync();
-      if (openMenu) {
-        (ev as BackButtonEvent).detail.register(MENU_BACK_BUTTON_PRIORITY, () => {
-          return openMenu.close();
-        });
-      }
-    });
-  }
+  doc?.addEventListener('ionBackButton', (ev: BackButtonEvent) => {
+    const openMenu = _getOpenSync();
+    if (openMenu) {
+      ev.detail.register(MENU_BACK_BUTTON_PRIORITY, () => {
+        return openMenu.close();
+      });
+    }
+  });
 
   return {
     registerAnimation,