+++ packages/angular/src/directives/control-value-accessors/select-value-accessor.ts
@@ -4,7 +4,7 @@ import { ValueAccessor } from '@ionic/angular/common';
 
 @Directive({
   /* tslint:disable-next-line:directive-selector */
-  selector: 'ion-range, ion-select, ion-radio-group, ion-segment, ion-datetime',
+  selector: 'ion-select, ion-radio-group, ion-segment, ion-datetime',
   providers: [
     {
       provide: NG_VALUE_ACCESSOR,
@@ -20,12 +20,7 @@ export class SelectValueAccessorDirective extends ValueAccessor {
 
   @HostListener('ionChange', ['$event.target'])
   _handleChangeEvent(
-    el:
-      | HTMLIonRangeElement
-      | HTMLIonSelectElement
-      | HTMLIonRadioGroupElement
-      | HTMLIonSegmentElement
-      | HTMLIonDatetimeElement
+    el: HTMLIonSelectElement | HTMLIonRadioGroupElement | HTMLIonSegmentElement | HTMLIonDatetimeElement
   ): void {
     this.handleValueChange(el, el.value);
   }